[inductor] Refactor fusion of inplace operations
Resubmit of #128979

`WeakDep`s force readers to have complete before a mutation overwrites the
buffer, but we want to allow fusions to occur for inplace mutations where the
same index is read and written.

Currently this is achieved by:
1. Identifying the buffers used by the mutating op in its `dep_closure`
2. Not creating `WeakDep`s for buffers in the `dep_closure`
3. Fixing up any bad fusions that might occur by an extra check in `can_fuse_vertical`

So we are first over-agressive in removing `WeakDep`, then add an ad-hoc fixup.

This PR instead emits all `WeakDep`s and adds a `fusable_weak_dep` check to
`can_fuse_vertical` which selectively allows inplace operation to fuse.

ghstack-source-id: fcfc4b146ff3f92c1cf348bdca5cebd72672b0cc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130835